# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiImageInput {
  prompt: string;
}

export interface GeminiImageOutput {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}

export interface CampaignSession {
  id: number;
  name: string;
  worldState: string;
  phase: string;
  /** @nullable */
  conversationId: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CampaignSessionInput {
  name: string;
  worldDescription?: string;
}

export interface CampaignSessionUpdate {
  worldState?: string;
  phase?: string;
}

export interface Player {
  id: number;
  sessionId: number;
  name: string;
  characterName: string;
  race: string;
  class: string;
  background: string;
  hp: number;
  maxHp: number;
  ac: number;
  level: number;
  stats: string;
  createdAt: string;
}

export interface PlayerInput {
  name: string;
  characterName: string;
  race: string;
  class: string;
  background: string;
  hp: number;
  maxHp: number;
  ac: number;
  level: number;
  stats: string;
}

export interface PlayerUpdate {
  hp?: number;
  maxHp?: number;
  ac?: number;
  stats?: string;
}

export interface DiceRollInput {
  playerId: number;
  diceType: string;
  result: number;
  purpose: string;
}

export interface DiceRollResult {
  id: number;
  playerId: number;
  diceType: string;
  result: number;
  purpose: string;
  createdAt: string;
}

export interface GmMessageInput {
  playerId: number;
  action: string;
  /** @nullable */
  diceRollId?: number | null;
}

export interface NarrativeEntry {
  id: number;
  sessionId: number;
  role: string;
  content: string;
  /** @nullable */
  playerId: number | null;
  createdAt: string;
}

export interface DiceRollEntry {
  id: number;
  playerId: number;
  characterName: string;
  diceType: string;
  result: number;
  purpose: string;
  createdAt: string;
}

export interface CampaignError {
  error: string;
}