# lib/api-zod/src/generated/types/campaignSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignSession {
  id: number;
  name: string;
  worldState: string;
  phase: string;
  /** @nullable */
  conversationId: number | null;
  createdAt: Date;
  updatedAt: Date;
}