# lib/api-zod/src/generated/types/diceRollEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiceRollEntry {
  id: number;
  playerId: number;
  characterName: string;
  diceType: string;
  result: number;
  purpose: string;
  createdAt: Date;
}