# lib/api-zod/src/generated/types/diceRollResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiceRollResult {
  id: number;
  playerId: number;
  diceType: string;
  result: number;
  purpose: string;
  createdAt: Date;
}